Some performance optimisations on the display standby script, but with no positive effect yet
Currently, the script uses ~4-5% of one cpu kernel and at times it goes
up to 100% and stays there. Don't know why but it drives me crazy. I
fixed a problem with a endless loop then the signal does not returns and
implemented the statistics pacakge for median method, because it's a
little faster than self made median calculation